+++ structs.go
@@ -15,10 +15,6 @@
 // This is because DBN encodes uint64 as strings, while the field annotations
 // know them as uint64.
 //
-// TODO
-//   * instrument definition
-//   * status
-//
 
 package dbn
 
@@ -888,4 +884,297 @@ func (r *StatMsg) Fill_Json(val *fastjson.Value, header *RHeader) error {
 	r.UpdateAction = uint8(val.GetUint("update_action"))
 	r.StatFlags = uint8(val.GetUint("stat_flags"))
 	return nil
+}
+
+///////////////////////////////////////////////////////////////////////////////
+
+// DataBento normalized Trading Status Update message.
+type StatusMsg struct {
+	Header                RHeader  `json:"hd" csv:"hd"`                                             // The record header.
+	TsRecv                uint64   `json:"ts_recv" csv:"ts_recv"`                                   // The capture-server-received timestamp expressed as number of nanoseconds since the UNIX epoch.
+	Action                uint16   `json:"action" csv:"action"`                                     // The type of status change.
+	Reason                uint16   `json:"reason" csv:"reason"`                                     // Additional details about the cause of the status change.
+	TradingEvent          uint16   `json:"trading_event" csv:"trading_event"`                       /// Further information about the status change and its effect on trading.
+	IsTrading             uint8    `json:"is_trading" csv:"is_trading"`                             /// The state of trading in the instrument.
+	IsQuoting             uint8    `json:"is_quoting" csv:"is_quoting"`                             /// The state of quoting in the instrument.
+	IsShortSellRestricted uint8    `json:"is_short_sell_restricted" csv:"is_short_sell_restricted"` // The state of short sell restrictions for the instrument.
+	Reserved              [7]uint8 // Filler for alignment.
+}
+
+const StatusMsg_Size = RHeader_Size + 24 // TODO check size, add test
+
+func (*StatusMsg) RType() RType {
+	return RType_Status
+}
+
+func (*StatusMsg) RSize() uint16 {
+	return StatusMsg_Size
+}
+
+func (r *StatusMsg) Fill_Raw(b []byte) error {
+	if len(b) < StatusMsg_Size {
+		return unexpectedBytesError(len(b), StatusMsg_Size)
+	}
+	err := r.Header.Fill_Raw(b[0:RHeader_Size])
+	if err != nil {
+		return err
+	}
+	body := b[RHeader_Size:] // slice of just the body
+	r.TsRecv = binary.LittleEndian.Uint64(body[0:8])
+	r.Action = binary.LittleEndian.Uint16(body[8:9])
+	r.Reason = binary.LittleEndian.Uint16(body[9:10])
+	r.TradingEvent = binary.LittleEndian.Uint16(body[10:11])
+	r.IsTrading = body[11]
+	r.IsQuoting = body[12]
+	r.IsShortSellRestricted = body[13]
+	return nil
+}
+
+func (r *StatusMsg) Fill_Json(val *fastjson.Value, header *RHeader) error {
+	r.Header = *header
+	r.TsRecv = fastjson_GetUint64FromString(val, "ts_recv")
+	r.Action = uint16(val.GetUint("action"))
+	r.Reason = uint16(val.GetUint("reason"))
+	r.TradingEvent = uint16(val.GetUint("trading_event"))
+	r.IsTrading = uint8(val.GetUint("is_trading"))
+	r.IsQuoting = uint8(val.GetUint("is_quoting"))
+	r.IsShortSellRestricted = uint8(val.GetUint("is_short_sell_restricted"))
+	return nil
+}
+
+///////////////////////////////////////////////////////////////////////////////
+
+// InstrumentDefMsg is a statistics message. A catchall for various data disseminated by publishers.
+// The [`stat_type`](Self::stat_type) indicates the statistic contained in the message.
+// This is not a strict byte-layout because RawSymbol has dynamic length that depends on metadata's SymbolCstrLen.
+type InstrumentDefMsg struct {
+	Header                  RHeader                        `json:"hd" csv:"hd"`                                                 // The common header.
+	TsRecv                  uint64                         `json:"ts_recv" csv:"ts_recv"`                                       // The capture-server-received timestamp expressed as the number of nanoseconds since the UNIX epoch.
+	MinPriceIncrement       int64                          `json:"min_price_increment" csv:"min_price_increment"`               // Fixed price The minimum constant tick for the instrument in units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	DisplayFactor           int64                          `json:"display_factor" csv:"display_factor"`                         // The multiplier to convert the venue’s display price to the conventional price, in units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	Expiration              uint64                         `json:"expiration" csv:"expiration"`                                 // The last eligible trade time expressed as a number of nanoseconds since the UNIX epoch. Will be [`crate::UNDEF_TIMESTAMP`] when null, such as for equities.
+	Activation              uint64                         `json:"activation" csv:"activation"`                                 // The time of instrument activation expressed as a number of nanoseconds since the UNIX epoch. Will be [`crate::UNDEF_TIMESTAMP`] when null, such as for equities.
+	HighLimitPrice          int64                          `json:"high_limit_price" csv:"high_limit_price"`                     // The allowable high limit price for the trading day in units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	LowLimitPrice           int64                          `json:"low_limit_price" csv:"low_limit_price"`                       // The allowable low limit price for the trading day in units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	MaxPriceVariation       int64                          `json:"max_price_variation" csv:"max_price_variation"`               // The differential value for price banding in units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	TradingReferencePrice   int64                          `json:"trading_reference_price" csv:"trading_reference_price"`       // The trading session settlement price on `trading_reference_date`.
+	UnitOfMeasureQty        int64                          `json:"unit_of_measure_qty" csv:"unit_of_measure_qty"`               // The contract size for each instrument, in combination with `unit_of_measure`, in units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	MinPriceIncrementAmount int64                          `json:"min_price_increment_amount" csv:"min_price_increment_amount"` // The value currently under development by the venue. Converted to units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	PriceRatio              int64                          `json:"price_ratio" csv:"price_ratio"`                               // The value used for price calculation in spread and leg pricing in units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	StrikePrice             int64                          `json:"strike_price" csv:"strike_price"`                             // The strike price of the option. Converted to units of 1e-9, i.e. 1/1,000,000,000 or 0.000000001.
+	InstAttribValue         int32                          `json:"inst_attrib_value" csv:"inst_attrib_value"`                   // A bitmap of instrument eligibility attributes.
+	UnderlyingID            uint32                         `json:"underlying_id" csv:"underlying_id"`                           // The `instrument_id` of the first underlying instrument.
+	RawInstrumentID         uint32                         `json:"raw_instrument_id" csv:"raw_instrument_id"`                   // The instrument ID assigned by the publisher. May be the same as `instrument_id`.
+	MarketDepthImplied      int32                          `json:"market_depth_implied" csv:"market_depth_implied"`             // The implied book depth on the price level data feed.
+	MarketDepth             int32                          `json:"market_depth" csv:"market_depth"`                             // The (outright) book depth on the price level data feed.
+	MarketSegmentID         uint32                         `json:"market_segment_id" csv:"market_segment_id"`                   // The market segment of the instrument.
+	MaxTradeVol             uint32                         `json:"max_trade_vol" csv:"max_trade_vol"`                           // The maximum trading volume for the instrument.
+	MinLotSize              int32                          `json:"min_lot_size" csv:"min_lot_size"`                             // The minimum order entry quantity for the instrument.
+	MinLotSizeBlock         int32                          `json:"min_lot_size_block" csv:"min_lot_size_block"`                 // The minimum quantity required for a block trade of the instrument.
+	MinLotSizeRoundLot      int32                          `json:"min_lot_size_round_lot" csv:"min_lot_size_round_lot"`         // The minimum quantity required for a round lot of the instrument. Multiples of this quantity are also round lots.
+	MinTradeVol             uint32                         `json:"min_trade_vol" csv:"min_trade_vol"`                           // The minimum trading volume for the instrument.
+	ContractMultiplier      int32                          `json:"contract_multiplier" csv:"contract_multiplier"`               // The number of deliverables per instrument, i.e. peak days.
+	DecayQuantity           int32                          `json:"decay_quantity" csv:"decay_quantity"`                         // The quantity that a contract will decay daily, after `decay_start_date` has been reached.
+	OriginalContractSize    int32                          `json:"original_contract_size" csv:"original_contract_size"`         // The fixed contract value assigned to each instrument.
+	TradingReferenceDate    uint16                         `json:"trading_reference_date" csv:"trading_reference_date"`         // The trading session date corresponding to the settlement price in  `trading_reference_price`, in number of days since the UNIX epoch.
+	ApplID                  int16                          `json:"appl_id" csv:"appl_id"`                                       // The channel ID assigned at the venue.
+	MaturityYear            uint16                         `json:"maturity_year" csv:"maturity_year"`                           // The calendar year reflected in the instrument symbol.
+	DecayStartDate          uint16                         `json:"decay_start_date" csv:"decay_start_date"`                     // The date at which a contract will begin to decay.
+	ChannelID               uint16                         `json:"channel_id" csv:"channel_id"`                                 // The channel ID assigned by Databento as an incrementing integer starting at zero.
+	Currency                [4]byte                        `json:"currency" csv:"currency"`                                     // The currency used for price fields.
+	SettlCurrency           [4]byte                        `json:"settl_currency" csv:"settl_currency"`                         // The currency used for settlement, if different from `currency`.
+	Secsubtype              [6]byte                        `json:"secsubtype" csv:"secsubtype"`                                 // The strategy type of the spread.
+	RawSymbol               [MetadataV2_SymbolCstrLen]byte `json:"raw_symbol" csv:"raw_symbol"`                                 // The instrument raw symbol assigned by the publisher.
+	Group                   [21]byte                       `json:"group" csv:"group"`                                           // The security group code of the instrument.
+	Exchange                [5]byte                        `json:"exchange" csv:"exchange"`                                     // The exchange used to identify the instrument.
+	Asset                   [7]byte                        `json:"asset" csv:"asset"`                                           // The underlying asset code (product code) of the instrument.
+	Cfi                     [7]byte                        `json:"cfi" csv:"cfi"`                                               // The ISO standard instrument categorization code.
+	SecurityType            [7]byte                        `json:"security_type" csv:"security_type"`                           // The type of the instrument, e.g. FUT for future or future spread.
+	UnitOfMeasure           [31]byte                       `json:"unit_of_measure" csv:"unit_of_measure"`                       // The unit of measure for the instrument’s original contract size, e.g. USD or LBS.
+	Underlying              [21]byte                       `json:"underlying" csv:"underlying"`                                 // The symbol of the first underlying instrument.
+	StrikePriceCurrency     [4]byte                        `json:"strike_price_currency" csv:"strike_price_currency"`           // The currency of [`strike_price`](Self::strike_price).
+	InstrumentClass         byte                           `json:"instrument_class" csv:"instrument_class"`                     // The classification of the instrument.
+	MatchAlgorithm          byte                           `json:"match_algorithm" csv:"match_algorithm"`                       // The matching algorithm used for the instrument, typically **F**IFO.
+	MdSecurityTradingStatus uint8                          `json:"md_security_trading_status" csv:"md_security_trading_status"` // The current trading state of the instrument.
+	MainFraction            uint8                          `json:"main_fraction" csv:"main_fraction"`                           // The price denominator of the main fraction.
+	PriceDisplayFormat      uint8                          `json:"price_display_format" csv:"price_display_format"`             // The number of digits to the right of the tick mark, to display fractional prices.
+	SettlPrice_type         uint8                          `json:"settl_price_type" csv:"settl_price_type"`                     // The type indicators for the settlement price, as a bitmap.
+	SubFraction             uint8                          `json:"sub_fraction" csv:"sub_fraction"`                             // The price denominator of the sub fraction.
+	UnderlyingProduct       uint8                          `json:"underlying_product" csv:"underlying_product"`                 // The product complex of the instrument.
+	SecurityUpdateAction    byte                           `json:"security_update_action" csv:"security_update_action"`         // Indicates if the instrument definition has been added, modified, or deleted.
+	MaturityMonth           uint8                          `json:"maturity_month" csv:"maturity_month"`                         // The calendar month reflected in the instrument symbol.
+	MaturityDay             uint8                          `json:"maturity_day" csv:"maturity_day"`                             // The calendar day reflected in the instrument symbol, or 0.
+	MaturityWeek            uint8                          `json:"maturity_week" csv:"maturity_week"`                           // The calendar week reflected in the instrument symbol, or 0.
+	UserDefinedInstrument   UserDefinedInstrument          `json:"user_defined_instrument" csv:"user_defined_instrument"`       // Indicates if the instrument is user defined: **Y**es or **N**o.
+	ContractMultiplierUnit  int8                           `json:"contract_multiplier_unit" csv:"contract_multiplier_unit"`     // The type of `contract_multiplier`. Either `1` for hours, or `2` for days.
+	FlowScheduleType        int8                           `json:"flow_schedule_type" csv:"flow_schedule_type"`                 // The schedule for delivering electricity.
+	TickRule                uint8                          `json:"tick_rule" csv:"tick_rule"`                                   // The tick rule of the spread.
+	Reserved                [10]byte                       `json:"_reserved" csv:"_reserved"`                                   // Filler for alignment.
+}
+
+// Minimum size of InstrumentDefMsg, the size with 0-length c-strings
+// We add 1*SymbolCstrLength to it to get actual size.
+// const InstrumentDefMsg_MinSize = RHeader_Size + 384 // TODO: count
+//
+//	func (*InstrumentDefMsg) RSize(cstrLength uint16) uint16 {
+//		return InstrumentDefMsg_MinSize + 2*cstrLength
+//	}
+
+const InstrumentDefMsg_Size = RHeader_Size + 313 + MetadataV2_SymbolCstrLen
+
+func (*InstrumentDefMsg) RType() RType {
+	return RType_InstrumentDef
+}
+
+func (*InstrumentDefMsg) RSize() uint16 {
+	return InstrumentDefMsg_Size
+}
+
+// func (r *InstrumentDefMsg) Fill_Raw(b []byte, cstrLength uint16) error {
+func (r *InstrumentDefMsg) Fill_Raw(b []byte) error {
+	if len(b) < StatMsg_Size {
+		return unexpectedBytesError(len(b), StatMsg_Size)
+	}
+	err := r.Header.Fill_Raw(b[0:RHeader_Size])
+	if err != nil {
+		return err
+	}
+	body := b[RHeader_Size:] // slice of just the body
+	r.TsRecv = binary.LittleEndian.Uint64(body[0:8])
+	r.MinPriceIncrement = int64(binary.LittleEndian.Uint64(body[8:16]))
+	r.DisplayFactor = int64(binary.LittleEndian.Uint64(body[16:24]))
+	r.Expiration = binary.LittleEndian.Uint64(body[24:32])
+	r.Activation = binary.LittleEndian.Uint64(body[32:40])
+	r.HighLimitPrice = int64(binary.LittleEndian.Uint64(body[40:48]))
+	r.LowLimitPrice = int64(binary.LittleEndian.Uint64(body[48:56]))
+	r.MaxPriceVariation = int64(binary.LittleEndian.Uint64(body[56:64]))
+	r.TradingReferencePrice = int64(binary.LittleEndian.Uint64(body[64:72]))
+	r.UnitOfMeasureQty = int64(binary.LittleEndian.Uint64(body[72:80]))
+	r.MinPriceIncrementAmount = int64(binary.LittleEndian.Uint64(body[80:88]))
+	r.PriceRatio = int64(binary.LittleEndian.Uint64(body[88:96]))
+	r.StrikePrice = int64(binary.LittleEndian.Uint64(body[96:104]))
+	r.InstAttribValue = int32(binary.LittleEndian.Uint32(body[104:108]))
+	r.UnderlyingID = binary.LittleEndian.Uint32(body[108:112])
+	r.RawInstrumentID = binary.LittleEndian.Uint32(body[112:116])
+	r.MarketDepthImplied = int32(binary.LittleEndian.Uint32(body[116:120]))
+	r.MarketDepth = int32(binary.LittleEndian.Uint32(body[120:124]))
+	r.MarketSegmentID = binary.LittleEndian.Uint32(body[124:128])
+	r.MaxTradeVol = binary.LittleEndian.Uint32(body[128:132])
+	r.MinLotSize = int32(binary.LittleEndian.Uint32(body[132:136]))
+	r.MinLotSizeBlock = int32(binary.LittleEndian.Uint32(body[136:140]))
+	r.MinLotSizeRoundLot = int32(binary.LittleEndian.Uint32(body[140:144]))
+	r.MinTradeVol = binary.LittleEndian.Uint32(body[144:148])
+	r.ContractMultiplier = int32(binary.LittleEndian.Uint32(body[148:152]))
+	r.DecayQuantity = int32(binary.LittleEndian.Uint32(body[152:156]))
+	r.OriginalContractSize = int32(binary.LittleEndian.Uint32(body[156:160]))
+	r.TradingReferenceDate = binary.LittleEndian.Uint16(body[160:162])
+	r.ApplID = int16(binary.LittleEndian.Uint16(body[162:164]))
+	r.MaturityYear = binary.LittleEndian.Uint16(body[164:166])
+	r.DecayStartDate = binary.LittleEndian.Uint16(body[166:168])
+	r.ChannelID = binary.LittleEndian.Uint16(body[168:170])
+	copy(r.Currency[:], body[170:174])            // byte[4]
+	copy(r.SettlCurrency[:], body[174:178])       // byte[4]
+	copy(r.Secsubtype[:], body[178:184])          // byte[6]
+	copy(r.RawSymbol[:], body[184:255])           // byte[MetadataV2_SymbolCstrLen] = 71
+	copy(r.Group[:], body[255:276])               // byte[21]
+	copy(r.Exchange[:], body[276:281])            // byte[5]
+	copy(r.Asset[:], body[281:288])               // byte[7]
+	copy(r.Cfi[:], body[288:295])                 // byte[7]
+	copy(r.SecurityType[:], body[295:302])        // byte[7]
+	copy(r.UnitOfMeasure[:], body[302:333])       // byte[31]
+	copy(r.Underlying[:], body[333:354])          // byte[21]
+	copy(r.StrikePriceCurrency[:], body[354:358]) // byte[4]
+	r.InstrumentClass = body[358]
+	r.MatchAlgorithm = body[359]
+	r.MdSecurityTradingStatus = body[360]
+	r.MainFraction = body[361]
+	r.PriceDisplayFormat = body[362]
+	r.SettlPrice_type = body[363]
+	r.SubFraction = body[364]
+	r.UnderlyingProduct = body[365]
+	r.SecurityUpdateAction = body[366]
+	r.MaturityMonth = body[367]
+	r.MaturityDay = body[368]
+	r.MaturityWeek = body[369]
+	r.UserDefinedInstrument = UserDefinedInstrument(body[370])
+	r.ContractMultiplierUnit = int8(body[371])
+	r.FlowScheduleType = int8(body[372])
+	r.TickRule = body[373]
+	return nil
+}
+
+func (r *InstrumentDefMsg) Fill_Json(val *fastjson.Value, header *RHeader) error {
+	r.Header = *header
+	r.TsRecv = fastjson_GetUint64FromString(val, "ts_recv")
+	r.MinPriceIncrement = fastjson_GetInt64FromString(val, "min_price_increment")
+	r.DisplayFactor = fastjson_GetInt64FromString(val, "display_factor")
+	r.Expiration = fastjson_GetUint64FromString(val, "expiration")
+	r.Activation = fastjson_GetUint64FromString(val, "activation")
+	r.HighLimitPrice = fastjson_GetInt64FromString(val, "high_limit_price")
+	r.LowLimitPrice = fastjson_GetInt64FromString(val, "low_limit_price")
+	r.MaxPriceVariation = fastjson_GetInt64FromString(val, "max_price_variation")
+	r.TradingReferencePrice = fastjson_GetInt64FromString(val, "trading_reference_price")
+	r.UnitOfMeasureQty = fastjson_GetInt64FromString(val, "unit_of_measure_qty")
+	r.MinPriceIncrementAmount = fastjson_GetInt64FromString(val, "min_price_increment_amount")
+	r.PriceRatio = fastjson_GetInt64FromString(val, "price_ratio")
+	r.StrikePrice = fastjson_GetInt64FromString(val, "strike_price")
+	r.InstAttribValue = int32(val.GetUint("inst_attrib_value"))
+	r.UnderlyingID = uint32(val.GetUint("underlying_id"))
+	r.RawInstrumentID = uint32(val.GetUint("raw_instrument_id"))
+	r.MarketDepthImplied = int32(val.GetUint("market_depth_implied"))
+	r.MarketDepth = int32(val.GetUint("market_depth"))
+	r.MarketSegmentID = uint32(val.GetUint("market_segment_id"))
+	r.MaxTradeVol = uint32(val.GetUint("max_trade_vol"))
+	r.MinLotSize = int32(val.GetUint("min_lot_size"))
+	r.MinLotSizeBlock = int32(val.GetUint("min_lot_size_block"))
+	r.MinLotSizeRoundLot = int32(val.GetUint("min_lot_size_round_lot"))
+	r.MinTradeVol = uint32(val.GetUint("min_trade_vol"))
+	r.ContractMultiplier = int32(val.GetUint("contract_multiplier"))
+	r.DecayQuantity = int32(val.GetUint("decay_quantity"))
+	r.OriginalContractSize = int32(val.GetUint("original_contract_size"))
+	r.TradingReferenceDate = uint16(val.GetUint("trading_reference_date"))
+	r.ApplID = int16(val.GetUint("appl_id"))
+	r.MaturityYear = uint16(val.GetUint("maturity_year"))
+	r.DecayStartDate = uint16(val.GetUint("decay_start_date"))
+	r.ChannelID = uint16(val.GetUint("channel_id"))
+	copy(r.Currency[:], val.GetStringBytes("currency"))
+	copy(r.SettlCurrency[:], val.GetStringBytes("settl_currency"))
+	copy(r.Secsubtype[:], val.GetStringBytes("secsubtype"))
+	copy(r.RawSymbol[:], val.GetStringBytes("raw_symbol"))
+	copy(r.Group[:], val.GetStringBytes("group"))
+	copy(r.Exchange[:], val.GetStringBytes("exchange"))
+	copy(r.Asset[:], val.GetStringBytes("asset"))
+	copy(r.Cfi[:], val.GetStringBytes("cfi"))
+	copy(r.SecurityType[:], val.GetStringBytes("security_type"))
+	copy(r.UnitOfMeasure[:], val.GetStringBytes("unit_of_measure"))
+	copy(r.Underlying[:], val.GetStringBytes("underlying"))
+	copy(r.StrikePriceCurrency[:], val.GetStringBytes("strike_price_currency"))
+	r.InstrumentClass = byte(val.GetUint("instrument_class"))
+	r.MatchAlgorithm = byte(val.GetUint("match_algorithm"))
+	r.MdSecurityTradingStatus = uint8(val.GetUint("md_security_trading_status"))
+	r.MainFraction = uint8(val.GetUint("main_fraction"))
+	r.PriceDisplayFormat = uint8(val.GetUint("price_display_format"))
+	r.SettlPrice_type = uint8(val.GetUint("settl_price_type"))
+	r.SubFraction = uint8(val.GetUint("sub_fraction"))
+	r.UnderlyingProduct = uint8(val.GetUint("underlying_product"))
+	r.SecurityUpdateAction = byte(val.GetUint("security_update_action"))
+	r.MaturityMonth = uint8(val.GetUint("maturity_month"))
+	r.MaturityDay = uint8(val.GetUint("maturity_day"))
+	r.MaturityWeek = uint8(val.GetUint("maturity_week"))
+	r.UserDefinedInstrument = UserDefinedInstrument(val.GetUint("user_defined_instrument"))
+	r.ContractMultiplierUnit = int8(val.GetUint("contract_multiplier_unit"))
+	r.FlowScheduleType = int8(val.GetUint("flow_schedule_type"))
+	r.TickRule = uint8(val.GetUint("tick_rule"))
+	return nil
+}
+
+func fillBytesFromStringValue(t []byte, val *fastjson.Value, key string) {
+	b := val.GetStringBytes(key)
+	for i := 0; i < len(t); i++ {
+		if i < len(b) {
+			t[i] = b[i]
+		} else {
+			t[i] = 0
+		}
+	}
 }